+++ core/src/main/java/com/github/karsaig/approvalcrest/matcher/CustomisableMatcher.java
@@ -11,7 +11,7 @@ package com.github.karsaig.approvalcrest.matcher;
 
 import org.hamcrest.Matcher;
 
-import com.google.gson.GsonConfiguration;
+import com.github.karsaig.json.GsonConfiguration;
 
 /**
  * {@link Matcher} implementation where fields and object types can be skipped from the comparison, or matched with
