+++ providers/gson/x/META-INF/services/com.github.karsaig.json.GsonBuilder
