+++ spi/src/main/java/com/github/karsaig/json/Gson.java
@@ -1,4 +1,4 @@
-package com.google.gson;
+package com.github.karsaig.json;
 
 /**
  * Very limited Json functionality that we need.
