+++ spi/src/main/java/com/github/karsaig/json/JsonElement.java
@@ -1,4 +1,4 @@
-package com.google.gson;
+package com.github.karsaig.json;
 
 public interface JsonElement {
     boolean isJsonArray();
