+++ spi/src/main/java/com/github/karsaig/json/JsonParserProvider.java
@@ -1,4 +1,4 @@
-package com.google.gson;
+package com.github.karsaig.json;
 
 import org.jetbrains.annotations.NotNull;
