+++ spi/src/main/java/com/github/karsaig/json/JsonSerializer.java
@@ -1,4 +1,4 @@
-package com.google.gson;
+package com.github.karsaig.json;
 
 import java.lang.reflect.Type;
