+++ spi/src/main/java/com/github/karsaig/json/graph/GraphAdapterBuilderProvider.java
@@ -1,4 +1,4 @@
-package com.google.gson.graph;
+package com.github.karsaig.json.graph;
 
 import org.jetbrains.annotations.NotNull;
